Added settings to navbar and changing password through admin works

+++ app/controllers/runners_controller.rb
@@ -53,9 +53,10 @@ class RunnersController < ApplicationController
   # PATCH/PUT /runners/1
   # PATCH/PUT /runners/1.json
   def update
+    byebug
     #Change Runs if group is updated on runner
     #Tests current group_id to group_id param passed in
-    if @runner.group_id.to_s != runner_params[:group_id]
+    if @runner.group_id.to_s != runner_params[:group_id] && runner_params[:group_id] != nil
       Run.where(runner_id: @runner.id).find_each do |oldRun|
         #Must destroy old runs for this runner
         oldRun.destroy
@@ -76,7 +77,7 @@ class RunnersController < ApplicationController
     respond_to do |format|
       if @runner.update(runner_params)
 
-        format.html { redirect_to runner_today_path(@runner), notice: 'Runner was successfully updated.' }
+        format.html { redirect_to today_path, notice: 'Runner was successfully updated.' }
         format.json { render :show, status: :ok, location: @runner }
       else
         format.html { render :edit }
